rename redash.scheduler to redash.schedule
